add vertices widget
